+++ src/main/gen/icu/windea/pls/script/psi/ParadoxScriptString.java
@@ -15,9 +15,6 @@ public interface ParadoxScriptString extends ParadoxScriptValue, PsiLiteralValue
   @Nullable
   ParadoxScriptParameter getParameter();
 
-  @Nullable
-  ParadoxScriptValueStringTemplate getValueStringTemplate();
-
   @NotNull
   String getValue();