+++ src/main/kotlin/icu/windea/pls/script/psi/IParadoxScriptStringTemplate.kt
@@ -1,5 +1,0 @@
-package icu.windea.pls.script.psi
-
-import com.intellij.psi.*
-
-interface IParadoxScriptStringTemplate: PsiElement