+++ src/main/kotlin/icu/windea/pls/script/psi/ParadoxScriptPsiExtensions.kt
@@ -143,7 +143,6 @@ fun PsiElement.findParentDefinitionProperty(fromParentBlock: Boolean = false): P
 	return null
 }
 
-//TODO 使用统一的缓存信息
 
 fun ParadoxScriptPropertyKey.isSimpleExpression():Boolean{
 	val singleChild = this.firstChild?.takeIf { it.nextSibling == null } ?: return false
@@ -153,12 +152,4 @@ fun ParadoxScriptPropertyKey.isSimpleExpression():Boolean{
 fun ParadoxScriptString.isSimpleExpression():Boolean{
 	val singleChild = this.firstChild?.takeIf { it.nextSibling == null } ?: return false
 	return singleChild.elementType.let { it == STRING_TOKEN || it == QUOTED_STRING_TOKEN }
-}
-
-/**
- * 判断keyExpression是否可能是linkExpression。（如：`root`、`root.owner`）
- */
-fun ParadoxScriptPropertyKey.isPossibleLinkExpression(): Boolean{
-	val propertyConfig = this.getPropertyConfig()  ?: return false
-	return CwtConfigHandler.supportsScopes(propertyConfig)
 }