+++ src/main/kotlin/icu/windea/pls/script/reference/ParadoxScriptValueExpressionReferenceProvider.kt
@@ -1,10 +1,11 @@
 package icu.windea.pls.script.reference
 
 import com.intellij.psi.*
-import com.intellij.psi.util.*
 import com.intellij.util.*
 import icu.windea.pls.*
+import icu.windea.pls.core.*
 import icu.windea.pls.script.psi.*
+import icu.windea.pls.script.psi.impl.*
 
 /**
  * @see ParadoxScriptValueReference
@@ -14,13 +15,14 @@ import icu.windea.pls.script.psi.*
 class ParadoxScriptValueExpressionReferenceProvider: PsiReferenceProvider(){
 	override fun getReferencesByElement(element: PsiElement, context: ProcessingContext): Array<PsiReference> {
 		if(element !is ParadoxScriptString) return PsiReference.EMPTY_ARRAY
-		element.processChild {
-			when(it.elementType) {
-				ParadoxScriptElementTypes.STRING_TOKEN -> return arrayOf(ParadoxScriptValueReference(element, it.textRangeInParent))
-				ParadoxScriptElementTypes.QUOTED_STRING_TOKEN -> return arrayOf(ParadoxScriptValueReference(element, it.textRangeInParent))
-				else -> end()
-			}
+		val expressionInfo = element.kvExpressionInfo ?: return PsiReference.EMPTY_ARRAY
+		return when(expressionInfo.type){
+			ParadoxKvExpressionType.LiteralType -> arrayOf(ParadoxScriptValueReference(element, expressionInfo.wholeRange))
+			ParadoxKvExpressionType.ParameterType -> PsiReference.EMPTY_ARRAY
+			ParadoxKvExpressionType.StringTemplateType -> PsiReference.EMPTY_ARRAY
+			ParadoxKvExpressionType.ScopeExpression -> expressionInfo.ranges.mapToArray { ParadoxScriptLinkReference(element, it) }
+			ParadoxKvExpressionType.ScopeValueExpression -> PsiReference.EMPTY_ARRAY //TODO
+			ParadoxKvExpressionType.ScriptValueExpression -> PsiReference.EMPTY_ARRAY //unexpected
 		}
-		return PsiReference.EMPTY_ARRAY
 	}
 }